Deduce basic conditional
Say we have a scenario where we have the right key pressed for the first
half. Additionally, the state reflects moving to the right for the first
half.

We determine that we can increase the x to move to the right for the
first half. We can also determine that the right button being down
coincides with the state moving to the right.

This is a basic implementation of this behaviour. Currently it requires
a perfect match, but it's worth adding some tolerance.